Fixing this by checking if staged download exists first

+++ coastal-hazards-portal/src/main/java/gov/usgs/cida/coastalhazards/rest/data/DownloadResource.java
@@ -208,6 +208,10 @@ public class DownloadResource {
 	public Response deleteStagedItem(@PathParam("id") String itemId, @Context HttpServletRequest request) {
 		Response response = null;
 		try (DownloadService downloadService = new DownloadService()) {
+			Download download = downloadService.get(itemId);
+			if (download == null) {
+				throw new NotFoundException();
+			}
 			boolean deleted = downloadService.delete(itemId);
 			response = Response.ok("{\"deleted\":\"" + deleted + "\"}", MediaType.APPLICATION_JSON_TYPE).build();
 		}
